backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
